requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
djangorestframework-files==1.1.0
djangorestframework-jsonapi==3.0.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
packaging==19.2
psycopg2-binary==2.8.4
pyparsing==2.4.5
pytz==2019.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.7
